Updated listening

+++ Project Zomboid/src/me/zeus/Zomboid/Util/Calculator.java
@@ -3,13 +3,13 @@ package me.zeus.Zomboid.Util;
 
 public class Calculator {
     
-    int average;
+    double average;
     
     public Calculator(){
         
     }
     
-    public double calculateAverage(int a, int b){        
+    public double calculateAverage(double a, double b){        
         average = a/b;
         return average;
     }
